Extracted frame related logic in sprite.ts https://github.com/achobanov/Web-Game/issues/11
